タイトルシーンの作成中 タイマーがうまく作動していない タイトル画像の入手

+++ src/Character_Base.cpp
@@ -74,7 +74,7 @@ void CharacterBase::Move_Player(bool* m_check_move, Vector3* camera_rot, Vector3
 	XINPUT_STATE input;
 	// ゲームパッドの情報を丸ごと取得
 	//GetJoypadXInputState(DX_INPUT_PAD1, &input);
-	GetJoypadXInputState(pad_no, &input);
+	GetJoypadXInputState((int)pad_no, &input);
 	// 左スティックの値を設定
 	mov.x = input.ThumbLX;
 	mov.z = input.ThumbLY;
